Reorganized router to multistep form
Fixes #148
Fixes #149
