+++ src/Components/Routers/SolicitarPrestamoRoute.js
@@ -1,54 +1,15 @@
 import { Navigate, Outlet } from 'react-router-dom';
 import { useAuthContext } from '../../Contexts/authContext';
-import { LOGIN, INGRESAR_DATOS, VERIFICAR_CORREO, VERIFICAR_TELEFONO, TU_HISTORIAL } from '../../Config/Router/paths';
-import IngresarDatos from '../../Pages/CreateAccount/IngresaTusDatos';
-import VerificarCorreo from '../../Pages/CreateAccount/VerificacionCorreo';
-import VerificarTelefono from '../../Pages/CreateAccount/VerificacionTelefono';
-import TuHistorial from '../../Pages/SolicitarPrestamo/TuHistorial';
+import { LOGIN} from '../../Config/Router/paths';
+
 import React from 'react'
 
 export default function SolicitarPrestamoRoute() {
-    const { tokenExist,  verificationStep, } = useAuthContext();
+    const { tokenExist,} = useAuthContext();
     console.log('Solicitar Prestamo route');
 
     if (!tokenExist) {
         return <Navigate to={LOGIN} />;
-    }
-
-    if (tokenExist) {
-        switch (verificationStep) {
-            case 1:
-                return (
-                    <div>
-                        <Navigate to={INGRESAR_DATOS} />
-                        <IngresarDatos />
-                    </div>
-                );
-            case 2:
-                return (
-                    <div>
-                        <Navigate to={VERIFICAR_CORREO} />
-                        <VerificarCorreo />
-                    </div>
-                );
-            case 3:
-                return (
-                    <div>
-                        <Navigate to={VERIFICAR_TELEFONO} />
-                        <VerificarTelefono />
-                    </div>
-                );
-            case 4:
-                return (
-                    <div>
-                        <Navigate to={TU_HISTORIAL} />
-                        <TuHistorial />
-                    </div>
-                );
-            default:
-                console.log('Verification Route default');
-                return null;
-        }
     }
