+++ src/Components/SolicitarPrestamo/VerificacionTelefono.js
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import axios from 'axios';
-
-import { Input, Button } from "@nextui-org/react";
+import { Input, Button } from "@nextui-org/react"
 import toast, { Toaster } from 'react-hot-toast';
 
-import NavbarLoan from '../../Components/Loan Service/NavbarLoan';
+import NavbarLoan from './NavbarLoan';
 import { useAuthContext } from '../../Contexts/authContext';
+
 
 const styles_input = {
     label: [
@@ -32,35 +32,34 @@ const styles_input = {
 
 axios.defaults.withCredentials = true;
 
-function VerificacionCorreo() {
-    const { tokenExist, checkToken, navigateToNextStep } = useAuthContext();
+function VerificacionTelefono() {
+    const { navigateToNextStep } = useAuthContext();
     const [otpCode, setOtpCode] = useState('');
 
     const handleSubmit = async () => {
-        //console.log( otpCode);
         try {
-            const response = await axios.post('https://bursapi.com/usuarios/verifyEmail', {
+            const response = await axios.post('https://bursapi.com/usuarios/verifyPhoneNumber', {
                 code: otpCode
             });
-            if (response.data.status === 'success') {
-                toast.success('Correo verificado con éxito');
+            if (response.data.message === 'Telefono verificado con éxito') {
+                toast.success('Telefono verificado con éxito');
                 setTimeout(() => {
-                    navigateToNextStep(3);
+                    navigateToNextStep(4);
                 }, 2000);
             }
+
         } catch (error) {
-            if(error.response.status === 400) {
+            if (error.response.status === 400) {
                 toast.error('Codigo incorrecto')
-            }
-            else {
-                toast.error('Error al verificar el correo, intentalo de nuevo')
+            } else {
+                toast.error('Error al verificar el telefono, intentalo de nuevo')
             }
         }
     };
 
     const handleResend = async () => {
         try {
-            const response = await axios.post('https://bursapi.com/usuarios/resendOTPCodeEmail');
+            const response = await axios.post('https://bursapi.com/usuarios/resendOTPCodePhoneNumber');
             if (response.data.status === 'success') {
                 toast('Codigo reenviado')
             }
@@ -68,6 +67,7 @@ function VerificacionCorreo() {
             toast.error('Error al reenviar el codigo')
         }
     };
+
     return (
         <div>
             <NavbarLoan />
@@ -101,8 +101,8 @@ function VerificacionCorreo() {
                     </ol>
                     <div className='flex flex-col space-y-12'>
                         <div className='flex flex-col space-y-5'>
-                            <h1 className='font-rubik font-bold text-2xl text-dark-blue-950'>Te enviamos un codigo al correo XX XXXX XXXX</h1>
-                            <p className='w-3/4 font-rubik font-medium text-sm text-dark-blue-800'>Ingresa el codigo OTP que te enviamos por correo</p>
+                            <h1 className='font-rubik font-bold text-2xl text-dark-blue-950'>Te enviamos un codigo al telefono XX XXXX XXXX</h1>
+                            <p className='w-3/4 font-rubik font-medium text-sm text-dark-blue-800'>Ingresa el codigo OTP que te enviamos por mensaje</p>
                         </div>
                         <div className='flex-col space-y-3'>
                             <Input
@@ -121,7 +121,7 @@ function VerificacionCorreo() {
                                 className='px-0 font-rubik font-medium text-xs text-purple-heart-700 data-[hover=true]:bg-default/0'
                                 onClick={handleResend}
                             >
-                            Reenviar codigo
+                                Reenviar codigo
                             </Button>
                         </div>
                         <Button
@@ -129,17 +129,17 @@ function VerificacionCorreo() {
                             color='secondary'
                             onClick={handleSubmit}
                         >
-                            Verficar Correo
+                            Verificar Telefono
                         </Button>
                         <Toaster
                             position="top-center"
                             reverseOrder={false}
                         />
-                    </div> 
+                    </div>
                 </div>
             </div>
         </div>
-  )
+    )
 }
 
-export default VerificacionCorreo
+export default VerificacionTelefono
