+++ src/Config/Router/BrowserRouter.js
@@ -1,13 +1,10 @@
 import { createBrowserRouter } from 'react-router-dom';
 import {
     LOGIN, SIGNUP, QUIERO_INVERTIR,
-    INGRESAR_DATOS, VERIFICAR_CORREO,
-    VERIFICAR_TELEFONO, VERIFICAR,
-    SOLICITAR_PRESTAMO, TU_HISTORIAL
+    PRESTAMO, PRESTAMO_SOLICITUD
 } from "../../Config/Router/paths";
 
 import PublicRoute from "../../Components/Routers/PublicRoute";
-import VerificationRoute from '../../Components/Routers/VerificationRoute';
 import SolicitarPrestamoRoute from '../../Components/Routers/SolicitarPrestamoRoute';
 
 import Navbar from "../../Components/Navbar";
@@ -18,11 +15,7 @@ import QuieroInvertir from "../../Pages/QuieroInvertir";
 import Login from "../../Pages/Login";
 import SignUp from '../../Pages/CreateAccount';
 
-import IngresarDatos from '../../Pages/CreateAccount/IngresaTusDatos';
-import VerificarCorreo from '../../Pages/CreateAccount/VerificacionCorreo';
-import VerificarTelefono from '../../Pages/CreateAccount/VerificacionTelefono';
-
-import TuHistorial from '../../Pages/SolicitarPrestamo/TuHistorial';
+import Solicitar from '../../Pages/SolicitarPrestamo/Solicitar';
 
 export const router = createBrowserRouter([  
     {
@@ -60,31 +53,12 @@ export const router = createBrowserRouter([
         ]
     },
     {
-        path: VERIFICAR,
-        element: <VerificationRoute />,
-        children: [
-            {
-                path: INGRESAR_DATOS,
-                element: <IngresarDatos />
-            },
-            {
-                path: VERIFICAR_CORREO,
-                element: <VerificarCorreo />
-            },
-            {
-                path: VERIFICAR_TELEFONO,
-                element: <VerificarTelefono />
-            }
-        ]
-
-    },
-    {
-        path: SOLICITAR_PRESTAMO,
+        path: PRESTAMO,
         element: <SolicitarPrestamoRoute />,
         children: [
             {
-                path: TU_HISTORIAL,
-                element: <TuHistorial />
+                path: PRESTAMO_SOLICITUD,
+                element: <Solicitar />
             }
         ]
     },
