+++ src/Config/Router/paths.js
@@ -5,8 +5,13 @@ export const PRIVATE = '/private';
 export const LOGOUT = '/private/logout';
 
 export const QUIERO_INVERTIR = '/quiero-invertir';
-export const VERIFICAR = '/verificar';
-export const INGRESAR_DATOS = '/verificar/ingresar-datos';
+export const PRESTAMO = '/prestamo';
+export const PRESTAMO_SOLICITUD = '/prestamo/solicitud';
+
+
+
+
+
 export const VERIFICAR_CORREO = '/verificar/verificar-correo';
 export const VERIFICAR_TELEFONO = '/verificar/verificar-telefono';
